major reorganization for clarity
