+++ my-electron-app/src/components/file_management/file_manager.js
