+++ my-electron-app/src/components/img-scroll-drag/model-img-drag.js
