+++ my-electron-app/src/components/python-terminal/python-terminal-renderer.js
