+++ my-electron-app/src/main_window/main_window.html
@@ -19,7 +19,7 @@
                 <br/>
             </div>
         </div>
-        <script src="./model-img-drag.js"></script>
+        <script src="../components/img-scroll-drag/model-img-drag.js"></script>
         <div class="page-bottom-content">
             <div class="console-text-wrapper" id="console-text-wrapper">
                 <p class="console-text" id="console-text">Default Text</p>
@@ -29,7 +29,7 @@
                 <input type="text" class="console-input" id="console-input"/>
             </div>
         </div>
-        <script src="./python-shell-terminal.js"></script>
+        <script src="../components/python-terminal/python-terminal-renderer.js"></script>
     </div>
     
 </body>
